Make level lighting more harsh and consistent.
I think the SpaceWorld demo used a uniform type of lighting, before
they messed with it on Mario and other models in 1996. Most of the
models use a consistent type of lighting, with scale 4
(the ambient color is 4 times less than the diffuse color).

I've seen inaccuracies in some levels because of this, notably
WF and CCM, but others too, where the level lighting just wasn't
harsh enough.

When I make the PR, I'll be sure to provide some reference images too.