Bowser Source Skybox (#124)
* Accurate Source Bowser Skybox

Using original source cloud textures and more info of the bottom layer of the skybox and more 1:1 details

* Bowser Skybox 

Using original source cloud textures and more info of the bottom layer of the skybox and more 1:1 details.

* Bowser Skybox Source

Forgot to upload the actual last one. Using the source cloud textures and more info of the bottom layer of the skybox and more accurate to the upper half of the skybox.

* Bowser Skybox Source Remake

Recreation of the Bowser In Dark World 1995 skybox using the real source cloud textures.